+++ basic/views/layouts/main.php
@@ -48,7 +48,7 @@ $this->registerLinkTag(['rel' => 'icon', 'type' => 'image/x-icon', 'href' => Yii
             ['label' => 'Aircrafts', 'url' => ['/aircraft/index']],
             ['label' => 'Airports', 'url' => ['/airport/index']],
             ['label' => 'Routes', 'url' => ['/route/index']],
-            ['label' => 'Submit FPL', 'url' => ['/submitted-flight-plan/select-route']],
+            ['label' => 'Submit FPL', 'url' => ['/submitted-flight-plan/index']],
             Yii::$app->user->isGuest
                 ? ['label' => 'Login', 'url' => ['/site/login']]
                 : '<li class="nav-item">'
